Monster Game
As the monsters and the player move, compare their locations, and
determine if a monster has reached a player. If any of the monsters
reaches a player, the game should end and the player should be give a
"GAME OVER" notification. This was created Ron Bryan Tan of 3-ITD from
the University of Santo Tomas